.github,Makefile: Add golangci-lint as a tool dependency
Update the Makefile and introduce the golangci-lint target to ensure
that local/CI environments use the same version of the golangci-lint
tool.

Update the sanity workflow and ensure CI is running the `lint` Makefile
target, which runs that golangci-lint target on the repository path.

Because we're replacing the golangci-lint github action with running a
Makefile target, we need to cache the go build/module paths so we don't
increase the overall CI sanity workflow time as that
action was previously handling that for us under-the-hood.

Signed-off-by: timflannagan <timflannagan@gmail.com>