[EntitySource] Modify entity source interface
The "Get" function of the entity source did not return an error and instead
only a nil entity when its not found. This PR modifies the same, to make it
return an error.

Reason:
The method could be implmented with a client that fetches bundle contents from
the cluster. The error returned from the client needs to be propagated.

Signed-off-by: Varsha Prasad Narsing <varshaprasad96@gmail.com>